Decode: DecodeImageFile loads through IO callbacks
AWESOME feature in stb_image that lets use decode images in tandem with
PhysicsFS without having to load the entire contents of the file into a
huge buffer and pass that to stb_image. Through io callback loads,
stb_image requests chunks of memory (or to skip through). This way we
don't need a FILE pointer or to buffer the whole file. Sweet!.